+++ Sources/Parsing/EventState.swift
@@ -1,7 +1,0 @@
-extension PureYAML.Parsing {
-    struct EventState {
-        var lines: [Line]
-        var index: Int
-        var events: [Event]
-    }
-}