+++ Sources/Parsing/NodeKind.swift
@@ -1,0 +1,8 @@
+extension PureYAML.Parsing {
+    enum NodeKind: Equatable {
+        case alias
+        case mapping
+        case scalar
+        case sequence
+    }
+}