+++ Sources/Parsing/NodeResult.swift
@@ -1,0 +1,6 @@
+extension PureYAML.Parsing {
+    struct NodeResult: Equatable {
+        var kind: NodeKind
+        var endMark: Mark
+    }
+}